fix: isolate sandbox runtime crate
