+++ crates/sandbox/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "rara-sandbox"
+version.workspace = true
+edition.workspace = true
+license.workspace = true
+
+[dependencies]
+anyhow.workspace = true
+rara-config = { path = "../config" }
+url = "2.5"
+uuid = { version = "1.11", features = ["v4"] }
+
+[dev-dependencies]
+tempfile.workspace = true
